+++ ts/packages/memory/conversation/src/index.ts
@@ -4,6 +4,7 @@
 export * from "./podcast.js";
 export * from "./podcastMessage.js";
 export * from "./importPodcast.js";
+export * from "./memory.js";
 export * from "./conversationMemory.js";
 export * from "./emailMemory.js";
 export * from "./emailMessage.js";
