+++ ts/packages/memory/conversation/src/memory.ts
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import * as kp from "knowpro";
+import { TypeChatLanguageModel } from "typechat";
+
+export type FileSaveSettings = {
+    dirPath: string;
+    baseFileName: string;
+};
+
+export interface MemorySettings {
+    conversationSettings: kp.ConversationSettings;
+    languageModel: TypeChatLanguageModel;
+    queryTranslator?: kp.SearchQueryTranslator | undefined;
+    fileSaveSettings?: FileSaveSettings | undefined;
+}
